ClinicProject/requirements-minimal.txt
# Minimal requirements for basic functionality
Django==5.2.8
djangorestframework==3.16.1
django-cors-headers==4.9.0
django-q2==1.8.0
redis==3.5.3
twilio==9.8.5
requests==2.32.5
python-decouple==3.8
pillow==12.0.0

# Optional AI dependencies (install separately if needed)
# torch==2.9.0+cpu
# transformers==4.57.1
# sentence-transformers==5.1.2
# scikit-learn==1.7.2
# numpy==2.3.4